Fixed a crash of unknown cause
Fixed: Schwi crashing when scanning an unavailable feed in a specific way... or something. Honestly, no clue what the issue is. Might even be the library I used to scan feeds but I'd rather not work with this hot, boiling mess of a code anymore so this'll work until I rewrite this whole thing again. Whenever that happens.